+++ src/main/java/ThirdStep/domain/CarService.java
@@ -1,13 +1,13 @@
-package ThirdStep.services;
+package ThirdStep.domain;
 
-import ThirdStep.interfaces.MovingCondition;
+import ThirdStep.domain.strategy.MovingStrategy;
 public class CarService {
 
     private static final int MOVE_FORWARD_VALUE = 1;
     private static final int STAY_VALUE = 0;
 
-    public static int movedDistance(MovingCondition movingCondition) {
-        if (movingCondition.canMoveForward()) {
+    public static int movedDistance(MovingStrategy movingStrategy) {
+        if (movingStrategy.canMoveForward()) {
             return MOVE_FORWARD_VALUE;
         }
         return STAY_VALUE;
