+++ src/main/java/ThirdStep/domain/model/RecordByRound.java
@@ -1,4 +1,4 @@
-package ThirdStep.model;
+package ThirdStep.domain.model;
 
 import java.util.ArrayList;
 import java.util.List;
