+++ src/main/java/ThirdStep/domain/strategy/MovingStrategy.java
@@ -1,0 +1,5 @@
+package ThirdStep.domain.strategy;
+
+public interface MovingStrategy {
+    boolean canMoveForward();
+}
