--- src/main/java/ThirdStep/interfaces/MovingCondition.java
@@ -1,5 +1,0 @@
-package ThirdStep.interfaces;
-
-public interface MovingCondition {
-    boolean canMoveForward();
-}
