+++ src/main/java/ThirdStep/view/utils/TextPrintUtils.java
@@ -1,4 +1,4 @@
-package ThirdStep.utils;
+package ThirdStep.view.utils;
 
 public class TextPrintUtils {
